Moved car code to new module named Player
I tried to document the code as much as possible since it's a little hard to read

The main idea is now instead of having the player logic in game.js, the logic is written in a separate module. The logic is divided up into methods that are part of the Player object. These methods can be called accordingly in game.js. Hopefully, this pattern can be used for gun objects as well. We might end up renaming player to Car so that it is distinct from the module(s) for Gun behavior