Fixed the Merge
Hyde figured out how to get the Laser Gun's image to show up and I was able to get the laser to show up.

The functions were all deleted and moved into the gameScene Class.  It was also determined that var config {} and var game should stay at the bottom of the file.