Rocket moves straight,  LaserGun code removed.
Rocket variable name was changed from Rock to Rocket for clarity.
The Rocket currently uses thrust just to move itself forward until it hits the border of the world.

Code that calculates the Laser's angle and the debug lines have been removed, as well as the variables used by the Laser Gun.